Make sure logging is always configured
Before this, logging wasn't configured when using only the GreyNoise API
object instead of the CLI.
