+++ src/greynoise/api.py
@@ -9,8 +9,10 @@ import structlog
 
 from greynoise.__version__ import __version__
 from greynoise.exceptions import RateLimitError, RequestFailure
-from greynoise.util import load_config, validate_ip
+from greynoise.util import configure_logging, load_config, validate_ip
 
+if not structlog.is_configured():
+    configure_logging()
 LOGGER = structlog.get_logger()
