+++ src/greynoise/cli/__init__.py
@@ -1,35 +1,12 @@
 """GreyNoise command line Interface."""
 
-import logging
-import sys
-
 import click
 import structlog
 from click_default_group import DefaultGroup
 from click_repl import register_repl
 
 from greynoise.cli import subcommand
-
-
-def configure_logging():
-    """Configure logging."""
-    logging.basicConfig(stream=sys.stderr, format="%(message)s", level=logging.CRITICAL)
-    logging.getLogger("greynoise").setLevel(logging.WARNING)
-    structlog.configure(
-        processors=[
-            structlog.stdlib.add_logger_name,
-            structlog.stdlib.add_log_level,
-            structlog.stdlib.PositionalArgumentsFormatter(),
-            structlog.processors.TimeStamper(fmt="%Y-%m-%d %H:%M.%S"),
-            structlog.processors.StackInfoRenderer(),
-            structlog.processors.format_exc_info,
-            structlog.dev.ConsoleRenderer(),
-        ],
-        context_class=dict,
-        logger_factory=structlog.stdlib.LoggerFactory(),
-        wrapper_class=structlog.stdlib.BoundLogger,
-        cache_logger_on_first_use=True,
-    )
+from greynoise.util import configure_logging
 
 
 @click.group(
@@ -40,7 +17,8 @@ def configure_logging():
 )
 def main():
     """GreyNoise CLI."""
-    configure_logging()
+    if not structlog.is_configured():
+        configure_logging()
 
 
 SUBCOMMAND_FUNCTIONS = [
