+++ src/greynoise/util.py
@@ -1,7 +1,9 @@
 """Utility functions."""
 
+import logging
 import os
 import socket
+import sys
 
 import structlog
 from six.moves.configparser import ConfigParser
@@ -10,6 +12,27 @@ CONFIG_FILE = os.path.expanduser(os.path.join("~", ".config", "greynoise", "conf
 LOGGER = structlog.get_logger()
 
 DEFAULT_CONFIG = {"api_key": "", "timeout": 60}
+
+
+def configure_logging():
+    """Configure logging."""
+    logging.basicConfig(stream=sys.stderr, format="%(message)s", level=logging.CRITICAL)
+    logging.getLogger("greynoise").setLevel(logging.WARNING)
+    structlog.configure(
+        processors=[
+            structlog.stdlib.add_logger_name,
+            structlog.stdlib.add_log_level,
+            structlog.stdlib.PositionalArgumentsFormatter(),
+            structlog.processors.TimeStamper(fmt="%Y-%m-%d %H:%M.%S"),
+            structlog.processors.StackInfoRenderer(),
+            structlog.processors.format_exc_info,
+            structlog.dev.ConsoleRenderer(),
+        ],
+        context_class=dict,
+        logger_factory=structlog.stdlib.LoggerFactory(),
+        wrapper_class=structlog.stdlib.BoundLogger,
+        cache_logger_on_first_use=True,
+    )
 
 
 def load_config():
