--- tests/conftest.py
@@ -1,9 +1,0 @@
-import pytest
-
-from greynoise.cli import configure_logging
-
-
-@pytest.fixture(scope="session", autouse=True)
-def configure_structlog():
-    """Configure logging for test cases."""
-    configure_logging()
